requirements.txt
# Dive Service Management - Python Dependencies
# Pin to minor version ranges for reproducibility with patch-level flexibility.

# Web Framework
Flask==3.1.0
Werkzeug==3.1.3

# Database
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
mysqlclient==2.2.6

# Authentication & Security
Flask-Security-Too==5.7.0
argon2-cffi==23.1.0

# Forms
Flask-WTF==1.2.2
WTForms==3.1.2
email-validator==2.2.0

# API / Serialization
marshmallow==3.22.0

# Task Queue (full deployment)
celery==5.4.0
redis==5.2.1

# Task Queue (lightweight / Pi deployment)
huey==2.5.2

# PDF Generation
fpdf2==2.8.2

# Excel Export
openpyxl==3.1.5

# Configuration
python-dotenv==1.0.1

# WSGI Server
gunicorn==22.0.0

# Mail (optional, for future notification emails)
Flask-Mail==0.10.0

# CLI
click==8.1.7
